+++ src/main/java/com/jlgs/howmuchah/entity/GroupMemberId.java
@@ -2,6 +2,7 @@ package com.jlgs.howmuchah.entity;
 
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -12,6 +13,7 @@ import java.util.UUID;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class GroupMemberId implements Serializable {
 
     private UUID groupId;